x/omni/types/message_whitelisted_test.go
package types

import (
	"testing"

	"omni/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgWhitelisted_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgWhitelisted
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgWhitelisted{
				Authority: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgWhitelisted{
				Authority: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
